+++ src/core_modules/capture-core/components/Pages/New/RegistrationDataEntry/RegistrationDataEntry.epics.js
@@ -16,6 +16,7 @@ import {
 import { convertFormToClient, convertClientToServer } from '../../../../converters';
 import { FEATURETYPE } from '../../../../constants';
 import { buildUrlQueryString, shouldUseNewDashboard } from '../../../../utils/routing';
+import { convertCategoryOptionsToServer } from '../../../../converters/clientToServer';
 
 const convertFn = pipe(convertFormToClient, convertClientToServer);
 
@@ -61,6 +62,7 @@ const deriveEvents = ({
     orgUnitId,
     redirectToEnrollmentEventNew,
     redirectToStageId,
+    attributeCategoryOptions,
 }) => {
     // in case we have a program that does not have an incident date (occurredAt), such as Malaria case diagnosis,
     // we want the incident to default to enrollmentDate (enrolledAt)
@@ -78,7 +80,10 @@ const deriveEvents = ({
             const dateToUseInActiveStatus =
             reportDateToUseInActiveStatus === 'enrolledAt' ? enrolledAt : sanitizedOccurredAt;
             const dateToUseInScheduleStatus = generateScheduleDateByEnrollmentDate ? enrolledAt : sanitizedOccurredAt;
-
+            const eventAttributeCategoryOptions = {};
+            if (attributeCategoryOptions) {
+                eventAttributeCategoryOptions.attributeCategoryOptions = convertCategoryOptionsToServer(attributeCategoryOptions);
+            }
             const eventInfo =
               openAfterEnrollment
                   ?
@@ -98,6 +103,7 @@ const deriveEvents = ({
 
             return {
                 ...eventInfo,
+                ...eventAttributeCategoryOptions,
                 programStage,
                 program: programId,
                 orgUnit: orgUnitId,
@@ -150,8 +156,16 @@ export const startSavingNewTrackedEntityInstanceWithEnrollmentEpic: Epic = (
         map((action) => {
             const { currentSelections: { orgUnitId, programId }, formsValues, dataEntriesFieldsValue } = store.value;
             const { dataStore, userDataStore, temp } = store.value.useNewDashboard;
-            const { occurredAt, enrolledAt, geometry } =
-                dataEntriesFieldsValue['newPageDataEntryId-newEnrollment'] || {};
+            const fieldsValue = dataEntriesFieldsValue['newPageDataEntryId-newEnrollment'] || {};
+            const { occurredAt, enrolledAt, geometry } = fieldsValue;
+            const attributeCategoryOptionsId = 'attributeCategoryOptions';
+            const attributeCategoryOptions = Object.keys(fieldsValue)
+                .filter(key => key.startsWith(attributeCategoryOptionsId))
+                .reduce((acc, key) => {
+                    const categoryId = key.split('-')[1];
+                    acc[categoryId] = fieldsValue[key];
+                    return acc;
+                }, {});
             const { trackedEntityType, stages } = getTrackerProgramThrowIfNotFound(programId);
             const values = formsValues['newPageDataEntryId-newEnrollment'] || {};
             const stageWithOpenAfterEnrollment = getStageWithOpenAfterEnrollment(stages);
@@ -165,6 +179,7 @@ export const startSavingNewTrackedEntityInstanceWithEnrollmentEpic: Epic = (
                 orgUnitId,
                 redirectToEnrollmentEventNew,
                 redirectToStageId: stageWithOpenAfterEnrollment?.id,
+                attributeCategoryOptions,
             });
             const { formFoundation, teiId: trackedEntity, uid } = action.payload;
             const formServerValues = formFoundation?.convertValues(values, convertFn);